requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-openai
streamlit_chat
utils
moviepy
azure-cognitiveservices-speech
docx2txt
faiss-cpu
